Create README.md
Bridge Of Light is Earth 2.0. community, we are creating a system and culture based on Natural Law / Simplify Divine Law. Is a community in search of the ultimate fair system which liberates all equally and abundantly. The wealth of the planet in all its form belongs to all of us not to some of us. We were here to obliterate and set humanity free of that influence and control, in a world where Universal law to the Hight degree was being broken. It does make logical rational and perfect sense to focus on Uniting the majority under Natural law and creating an instantly new system. We are a bridge of light to others to find their way to the New Earth timeline. we are creating new parts, and building a new network.